refactor: DTO를 이너클래스로 변경, 컨트롤러에서 Board를 직접 사용하지 않도록 변경
